Added notes

+++ src/PointOfSale/Article.java
@@ -1,20 +1,24 @@
 package PointOfSale;
 
+/**
+ * The class that defines the article object.
+ */
+
 public class Article {
 
-    private int Id;
+    private long Id;
     private String name;
     private double cost;
 
-    public int getId() { return Id; }
+    public long getId() { return Id; }
     public String getName() { return name; }
     public double getCost() { return cost; }
 
-    public void setId(int Id) { this.Id = Id; }
+    public void setId(long Id) { this.Id = Id; }
     public void setName(String name) { this.name = name; }
     public void setCost(double cost) { this.cost = cost; }
 
-    public Article(int Id, String name, double cost) {
+    public Article(long Id, String name, double cost) {
         this.Id = Id;
         this.name = name;
         this.cost = cost;
